refactor: RepositoryProtocol 분리
