+++ suhyeon/ExchangeRateCalculation/ExchangeRateCalculation/Domain/Entity/ErrorType/AppError.swift
