+++ suhyeon/ExchangeRateCalculation/ExchangeRateCalculation/Domain/Entity/ErrorType/CalculatorError.swift
