+++ suhyeon/ExchangeRateCalculation/ExchangeRateCalculation/Domain/Entity/ErrorType/CoreDataError.swift
