+++ suhyeon/ExchangeRateCalculation/ExchangeRateCalculation/Domain/Entity/ErrorType/NetworkError.swift
